requirements.txt
fastapi==0.104.1
labelme==5.3.1
matplotlib==3.3.4
ml_collections==0.1.0
nets==0.0.3.1
numpy==1.24.4
opencv_python==4.7.0.68
opencv_python_headless==4.8.1.78
Pillow==10.1.0
Requests==2.31.0
scikit_image==0.19.3
scikit_learn==1.2.1
scipy==1.11.3
tensorboardX==2.6.2.2
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
uvicorn==0.24.0.post1
